Prevent entire empty card stack from converting to same card
Previously, if you right-clicked an ItemStack of blank cards,
applyCDWDtoStack would apply to the entire stack, making them
all the same (randomized) card.

With this PR, the stack will shed a single card to become the
newly-randomized card. The rest of the stack will remain blank.

In addition, I did a little style cleanup.