Visuals overhaul
The user now has the ability to specify the color of cards and packs. This is done in the configs with an RGB integer--either decimal or hex, depending on whether the user prefixes with a # symbol.

All card and pack graphics are now generated using just two textures: a grayscale that is colorized based on user specifications, and an overlay which is not colorized (usually an accent piece like a frame or a seal). Correspondingly, I've removed the older card and pack textures since they're no longer used.

Users can now also specify the color of edition or custom packs by including a "color": field in the json. The color entry must be an integer rgb code (either integer or hex using # like in the configs), and must be surrounded by "" because it is taken in as a string and interpreted from there.

If a user does not specify this color (like with already-existing packs), or their specification is malformed, the pack will default to a procedurally-generated color by hashing the pack's "id" field.

The item binder graphic has been updated too.

The texture for the Card Master has been changed to something less mundane. He has cards in his cap!

I've added specialized textures to the villagenames resource directory, so that users can take advantage of the modular skin system without having to provide their own texture via a resourcepack. They will still need to use the relevant values in their VN configs, though.

I've changed the Card House a bit, mostly to add an overhang to the roof because that looks nicer; and added a path block in front of the door (and cleared out an extra space in case the door is obstructed by an awkward road position). In addition, I found and addressed a mistake when setting biome-specific blocks into the building. I also slightly adjusted the default values for villager trades.

Lastly, I've rolled back the config ID to 1.1, because I figure that it doesn't need to change multiple times before being compiled.