Add update checker
I've added machinery that will determine whether the mod is using the most current version. But to do this I had to make some substantial changes to the repo.

I've massively restructured the project (i.e., the equivalent of going up a few folder levels) in order to include the build.gradle file, so that people will have an easier time building.

Another thing the build.gradle file does is that it automatically substitutes the MOD_VERSION variable with the contents of a file called CURRENT_VERSION.

=== TO UPDATE THE VERSION OF THE PROJECT FROM NOW ON, ONLY CHANGE THE CONTENTS OF "CURRENT_VERSION"!! ===

An added benefit here is that the update checker pings the contents of this file and compares it to the mod's version. If it's different, an update message will be displayed.

I've also updated mcmod.info with a simplified description, and I went ahead and added my name.