Update drops
I've refactored the mob drop event so that all potentially dropped items are tallied, and then dropped in oder of item type. This allows additional packs dropped by bosses to be bundled together.

To the configs, I've done a number of things, which will either add new config options, OR in some cases change what's there:
-Added full customization to the drops from the Ender Dragon, or other bosses.
-Changed the drop rates for packs to use floats instead of integers, so that someone could do something like 1 in 2.5 (40%).
-Added drop options for cards
-Renamed original drop options to specify that they're for packs
-Added option to consolidate random drops, so that entities will drop at most one item (rather than independently dropping multiple types)